Added VM and Default Mode
Summary:
Added ability to select a VM mode in the aurora.config file. In this
mode benchmarks have their throughput requirements etc reduced as well
as Aurora switches to using the in memory version of its checkpointing
system. File system benchmarks still require some form of a stripe

Found a misconfiguration and required to specifically set the frequency
of checkpoints within the filebench benchmarks.

Test Plan: Test on a VM

Reviewers: tavianator, etsal, ali

Reviewed By: etsal

Differential Revision: https://review.rcs.uwaterloo.ca/D570