properly configure Firefox and update README
Summary: In order to replicate the numbers for Firefox we need to run in single process mode, as we did for the original submissions. Since this new version of Firefox has a different resident set size, update the README to explain the discrepancy between the numbers.

Test Plan: Running table 6

Reviewers: tavianator, ryan, ali

Reviewed By: ali

Differential Revision: https://review.rcs.uwaterloo.ca/D608